refactor(capture): seperate capture functions into different files
