+++ lib/core/runtime/capture/index.ts
@@ -11,12 +11,13 @@ import * as http from "http";
 import * as https from "https";
 import { URL } from "url";
 import { Socket, isIP } from "net";
+import { captureRequestBody } from "./request";
+import { captureResponseBody, ResponseBodyInfo } from "./response";
 
 import currentContext from "../../context";
 import logger from "../../logger/index";
 
 type requestProtocol = "http:" | "https:";
-type chunk = Buffer | Uint8Array | string | null | any;
 
 interface Timestamps {
   timeStart: number;
@@ -30,14 +31,10 @@ interface RequestInfo {
   localPort?: string | number;
   requestBody?: string;
 }
-interface ResponseInfo {
+interface ResponseInfo extends ResponseBodyInfo {
   remoteAddress?: string;
   remotePort?: string | number;
-  responseBody?: chunk[];
-  contentLength?: number;
 }
-
-const maxBodySize = 512 * 1024;
 
 /**
  * Convert a URL instance to a http.request options
@@ -62,88 +59,6 @@ const urlToOptions = (url: URL): http.RequestOptions => {
   }
 
   return options;
-};
-
-const captureRequestBody = (request: http.ClientRequest): void => {
-  let bodySize = 0;
-  const body: Buffer[] = [];
-
-  (request as any)._send = ((fn) => (
-    data: Buffer | Uint8Array | string,
-    encodingOrCallback?: string | ((err?: Error) => void) | undefined,
-    callbackOrUndefined?: ((err?: Error) => void) | undefined
-  ): boolean => {
-    let encoding: BufferEncoding = null;
-    let callback: (err?: Error) => void;
-
-    if (typeof encodingOrCallback === "function") {
-      callback = encodingOrCallback;
-    } else if (typeof callbackOrUndefined === "function") {
-      encoding = encodingOrCallback as BufferEncoding;
-      callback = callbackOrUndefined;
-    }
-
-    const buffer = ((): Buffer => {
-      if (Buffer.isBuffer(data)) {
-        return data;
-      }
-
-      if (typeof data === "string") {
-        return Buffer.from(data, encoding);
-      }
-
-      return Buffer.from(data);
-    })();
-
-    bodySize += buffer.length;
-    if (bodySize < maxBodySize) {
-      body.push(buffer);
-    }
-
-    return fn.apply(request, [data, encoding, callback]);
-  })((request as any)._send);
-
-  (request as any)._finish = ((fn) => (
-    ...args: unknown[]
-  ): void => {
-    (request as any)._body = ((): Buffer => {
-      // 请求结束后，将_body转换为buffer
-      if (!Buffer.isBuffer(body)) {
-        return Buffer.concat(body);
-      }
-
-      return body;
-    })();
-
-    (request as any)._bodySize = bodySize;
-
-    return fn.apply(request, args);
-  })((request as any)._finish);
-};
-
-const captureResponseBody = (
-  response: http.IncomingMessage,
-  handler: (d: any) => void
-): void => {
-  const originPush = response.push;
-
-  let { head } = (response as any).readableBuffer;
-  while (head) {
-    handler(head.data);
-    head = head.next;
-  }
-
-  response.push = (chunk: any, encoding?: string): boolean => {
-    try {
-      if (chunk) {
-        handler(chunk);
-      }
-    } catch (error) {
-      logger.debug(`capture stream chunk error ${error.message}`);
-    }
-
-    return originPush.call(response, chunk, encoding);
-  };
 };
 
 export const hack = <T extends typeof http.request>(
@@ -188,7 +103,8 @@ export const hack = <T extends typeof http.request>(
 
     const requestInfo: RequestInfo = {};
     const responseInfo: ResponseInfo = {
-      contentLength: 0
+      body: [],
+      bodyLength: 0
     };
     const timeStart = Date.now();
     const timestamps: Timestamps = {
@@ -199,7 +115,7 @@ export const hack = <T extends typeof http.request>(
     // 请求结束后，将请求信息记录在logger.json中
     const recordInLogger = (): void => {
       logger.debug(`record request info -> request length: ${
-        responseInfo.contentLength}`);
+        responseInfo.bodyLength}`);
     };
 
     const finishRequest = (): void => {
@@ -271,7 +187,8 @@ export const hack = <T extends typeof http.request>(
     request.once("finish", () => {
       let requestBody: string;
       const length = (request as any)._bodySize;
-      if (length >= maxBodySize) {
+      const tooLarge = (request as any)._bodyTooLarge;
+      if (tooLarge) {
         requestBody = Buffer.from(`body was too large too show, length: ${
           length}`).toString("base64");
       } else {
@@ -299,14 +216,7 @@ export const hack = <T extends typeof http.request>(
         response.headers["content-encoding"]
       }`);
 
-      const handler = (chunk: chunk): void => {
-        responseInfo.contentLength += chunk.length;
-        if (responseInfo.contentLength <= maxBodySize) {
-          responseInfo.responseBody.push(chunk);
-        }
-      };
-
-      captureResponseBody(response, handler);
+      Object.assign(responseInfo, captureResponseBody(response));
 
       response.once("close", () => {
         logger.debug(`${logPre} close`);
@@ -317,7 +227,7 @@ export const hack = <T extends typeof http.request>(
         cost = Date.now() - timeStart;
 
         logger.debug(`${logPre} end size：${
-          responseInfo.contentLength
+          responseInfo.bodyLength
         }, receive data cost: ${cost}ms'`);
 
         finishRequest();
